auto_tweet_project/requirements.txt
# Django関連
Django>=5.1.0,<5.2.0
python-dotenv>=1.0.0

# Twitter API関連
tweepy>=4.14.0

# 画像処理
Pillow>=10.0.0

# スケジューラー
python-crontab>=2.7.1

# 開発ツール
django-debug-toolbar>=4.2.0 